backend/requirements.txt
onnxruntime
Pillow
flask
waitress
opencv-python
scikit-image
shapely==2.0.1
flask-swagger-generator
flask_swagger_ui
